requirements.txt
streamlit
groq
gTTS
streamlit-mic-recorder
Pillow